Дюкин Петр 5.3

+++ 05-java-oop-principles/task03/src/com/example/task03/Milliseconds.java
@@ -28,6 +28,6 @@ public class Milliseconds implements TimeUnit {
 
     @Override
     public long toHours() {
-        return Math.round(amount / (1000 * 60 * 60));
+        return Math.round(amount / (double) (1000 * 60 * 60));
     }
 }
